some clean ups and new plaintext modulus

+++ crates/fhe/src/bfv/keys/secret_key.rs
@@ -262,8 +262,8 @@ mod tests {
 
         sk.coeffs.iter().for_each(|ci| {
             // Check that this is a small polynomial
-            let sk_variance = params.variance / 20.0;
-            assert!((*ci).abs() <= 2.0 * sk_variance)
+            let sk_variance = params.variance as f32 / 20.0;
+            assert!((*ci).abs() as f32 <= 2.0 * sk_variance)
         })
     }
